NEW BATTLE FUNCTIONALITY 1: REORGANIZATION
This commit is the first step to re-doing battle functionality with all that I have learned. This is just re-organizing files right now, but it's a first step for the next branch which is going to be refactoring
and re-doing the model itself. The entities and battlescene mainly.
