+++ src/view/BattleButtonPanel.java
@@ -1,11 +1,10 @@
-package view.gui;
+package view;
 
 import model.ety.BattleChoice;
 
 import javax.swing.*;
 import javax.swing.border.Border;
 import javax.swing.border.CompoundBorder;
-import javax.swing.border.EmptyBorder;
 import java.awt.*;
 import java.awt.event.ActionListener;
 import java.util.ArrayList;
