+++ src/view/ContentPanel.java
@@ -1,4 +1,4 @@
-package view.gui;
+package view;
 
 import javax.swing.*;
 import java.awt.*;
